requirements.txt
numpy==1.21.1
scipy==1.7.1
pandas==1.3.1

# choose one
tensorflow-gpu==2.4.1
#tensorflow==2.4.1

tensorflow_datasets==4.4.0
tensorflow_addons==0.12.1

deephyper==0.2.5
pymoo==0.4.2.2

nltk==3.6.5
networkx==2.6.2

matplotlib==3.4.2